Refactor: replce FeedbackPriority table with enum status
