+++ src/main/kotlin/com/parkjunhyung/IryeokFitAi/domain/feedback/entity/ENUM/FeedbackPriority.kt
@@ -1,0 +1,7 @@
+package com.parkjunhyung.IryeokFitAi.domain.feedback.entity.ENUM
+
+enum class FeedbackPriority {
+    HIGH,
+    MEDIUM,
+    LOW,
+}
