--- src/main/kotlin/com/parkjunhyung/IryeokFitAi/domain/feedback/entity/FeedbackPriority.kt
@@ -1,16 +1,0 @@
-package com.parkjunhyung.IryeokFitAi.domain.feedback.entity
-
-import jakarta.persistence.*
-
-@Table(name = "feedback_priority")
-@Entity
-class FeedbackPriority (
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0L,
-
-    @Column(name = "level", nullable = false, length = 50)
-    val level: String,
-
-)
