--- src/main/kotlin/com/parkjunhyung/IryeokFitAi/domain/feedback/repository/FeedbackPriorityRepository.kt
@@ -1,9 +1,0 @@
-package com.parkjunhyung.IryeokFitAi.domain.feedback.repository
-
-import com.parkjunhyung.IryeokFitAi.domain.feedback.entity.FeedbackPriority
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface FeedbackPriorityRepository: JpaRepository<FeedbackPriority, Long> {
-}
